Removing NetworkedPlayer & object setup
