+++ CSC8508/NetworkedGame.cpp
@@ -4,15 +4,15 @@
 //
 
 #include "NetworkedGame.h"
-#include "NetworkPlayer.h"
-#include "NetworkObject.h"
 #include "GameServer.h"
-#include "GameClient.h"
 #include "RenderObject.h"
 #include "INetworkComponent.h"
 #include "INetworkDeltaComponent.h"
 #include "ComponentManager.h"
 #include "EventManager.h"
+#include "GameObject.h"
+#include "GameClient.h"
+
 
 #define COLLISION_MSG 30
 
@@ -317,19 +317,4 @@ void NetworkedGame::ReceivePacket(int type, GamePacket* payload, int source)
 	ReceiveSpawnPacket(type, payload);
 	if (thisClient) thisClient->ReceivePacket(type, payload, source);
 	else if (thisServer) thisServer->ReceivePacket(type, payload, source);
-}
-
-void NetworkedGame::OnPlayerCollision(NetworkPlayer* a, NetworkPlayer* b) {
-	if (thisServer) 
-	{ 
-		MessagePacket newPacket;
-		newPacket.messageID = COLLISION_MSG;
-		newPacket.playerID  = a->GetPlayerNum();
-
-		thisClient->SendPacket(newPacket);
-
-		newPacket.playerID = b->GetPlayerNum();
-		thisClient->SendPacket(newPacket);
-
-	}
 }
