+++ CSC8508CoreClasses/GameObject.cpp
@@ -9,6 +9,7 @@
 #include <unordered_map>
 #include <unordered_set>
 #include <typeindex>
+
 #include <queue>
 using namespace NCL::CSC8508;
